requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
python-multipart>=0.0.9
python-dotenv>=1.0.1